# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Moments API",
        "contact": {},
        "version": "0.2.1"
    },
    "paths": {
        "/api/comment/add": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "添加评论",
                "parameters": [
                    {
                        "description": "添加评论",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.AddCommentReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/comment/remove": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "删除评论",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "评论ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "登录TOKEN",
                        "name": "x-api-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/file/s3PreSigned": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tag"
                ],
                "summary": "标签列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录TOKEN",
                        "name": "x-api-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.tagListResp"
                        }
                    }
                }
            }
        },
        "/api/file/upload": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File"
                ],
                "summary": "上传图片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录TOKEN",
                        "name": "x-api-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.tagListResp": {
            "type": "object",
            "properties": {
                "tags": {
                    "description": "标签列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "vo.AddCommentReq": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "正文",
                    "type": "string"
                },
                "email": {
                    "description": "作者邮箱",
                    "type": "string"
                },
                "memoId": {
                    "description": "回复的memoID",
                    "type": "integer"
                },
                "replyTo": {
                    "description": "回复给某人",
                    "type": "string"
                },
                "token": {
                    "description": "google recaptcha的token,开启的话必填",
                    "type": "string"
                },
                "username": {
                    "description": "作者用户名,未登录时才有效",
                    "type": "string"
                },
                "website": {
                    "description": "作者网站",
                    "type": "string"
                }
            }
        },
...
    }
}